Client: rejoin only once every specified silent time
The previous logic would have tried rejoining again and again with very short
interval inbetween once the threshold of silence is crossed for the 1st time.

But what is wanted is that if there is silence for a given time, then try
rejoining the mcast group and then again wait for that given time before
retrying to join again. This has been achieved now.