Client: Fixed repeated mcast rejoin
THere was a minimally harmful bug in the mcast rejoin logic, which
was that every predefined interval it was rejoin the mcast channel
/ group irrespective of whether mcast data was available or not.
This could have lead to some packet loss, when it is in the middle
of the mcast rejoin, in some cases, depending on how nw devices
like switches, wifi-aps etc buffer & handle sending of mcast pkts
to old but new again clients in a mcast group, with a small time
window when such client will be outside the mcast group.

Now the client will rejoin only if no multicast data is seen, by
default. However through a compile time define MCASTREJOIN_ALWAYS,
one could force the logic to always reconnect after a predefined
interval.