Client: First step towards stateless client logic
Want to change the client logic to be stateless in that

To do the same

a) I now create both the mcast and ucast socket at the beginning itself.

b) It inturn calls into a generic / common run routine which handles
   different types of packets as required.

Have created a ctags file to help jump throu code easily

Notes updated with the idea behind the stateless client functioning.