StatelessClient: Manage State, initialise llLostPkts
Added a global state variable to help identify if the program is
running for the first time wrt current context or if it is already
successfuly done with the current context.

If the state is DO and LostPackets list node count is 0

* even before any packet is processed,

then it means the logic is handling this context for the 1st time.
So add a entry to llLostPkts which indicates that all blocks/packets
belonging to this transfer/test context is missing.

* after processing of the current packet,

then it means the logic has successfully handled all blocks belonging
to this transfer/test context. So the state is set to DONE.

A Generic and a Data specific pkt counter are maintained.