StatelessClient: Have uCtxtId in snm, identify new context
Have added uCtxtId to snm to maintain the current test/transfer
context uniquie identifier, which will be passed from server to
uniquely identify a given test/transfer context.

A test / transfer context could be fully handled in a single session
and or across multiple sessions on both server and or client side.

Lack of a valid context id is used to identify that a new context
is being started and that is inturn used to setup the llLostPkts
appropriately at the context's beginning session.