StatelessClient: Cleanup, mcast rejoin, save context at end
Removed the now unused mcast_recv, ucast_recover.

Now snm_run does mcast rejoin when required.

The program saves the context, at the end. Currently the code will
not reach here. But later it may, if I had a signal handler for say
SIGUSR1 or so, which clears the global gbSNMRun so that snm_run
returns.

Track the largest data seq number got till that point in time.