Client: Remove iMaxDataSeqNumProc, Save theSrvrPeer instead
The stateless logic no longer requires the distinction between the
MaxDataSeqNum packet actually Got AND MaxDataSeqNum associated with
the current transfer, which may not be got but which is one requires
to be aware of to handle recovery properly.

So MaxDataSeqNumProcessed has been removed. And in the context file
in its place now theSrvrPeer data is saved and restored as required.

Even handle pi logic now alerts if the packet recieved from a diff
peer server than what was previously know. Handle UR already does
the same.