Client: Remove runmodes vestiges, Remove BCastAddr
As broadcasting of PI is no longer done by the client (Now the
server sends a PIReq on a specified mcast addr and port and inturn
the client sends a ucast PIAck to the server.), it is no longer
got from the commandline, nor saved in the snm context.

The runmodes was still specified in the usage help, the same has
been removed.