Client: Save NwContextId, DataFile, ContextFileBase
Now when saving the snm context it saves the following

The Nw Context Id - THis can help distinguish betweeen different
nw tests / transfers and their corresponding context on the client
side.

The DataFile - The name of the data file being used is saved into
context. This inturn is indirectly linked to the Nw Context Id.
This can be used later to automatically switch to that data file
and continue a test / transfer, based on what nw context id is
recieved with the nw packets.

The ContextFileBase - The base part of the file name used for saving
the context is also saved in the context file. So if at a later date
auto loading of saved context based on Nw Context Id is implemented,
then any new context being saved corresponding to the same will be
written to the original location which was used for saving the original
context.

To ensure that one nw context doesn't overwrite client side context
corresponding to a different nw context id, now the nw context id
is also part of the context file name used for saving.