Client: Context AutoLoading skeleton
The logic used to determine the context file name from context file
base name, nw context id and tag has been moved into a new func from
the save_context function.

This function is inturn called by the save_context function.

Now context_load function takes a additional mode argument which tells
it whether it is being called at the beginning of the program or as
part of auto loading of context based on change in NwContext Id of
recieved packet.

If called at the beginning, and inturn if no Context File Name, then
no context is loaded. However if being called due to CTXTAUTOLOAD
logic, then it will determine the required context file name using the
new function specified at the beginning and then inturn load the context
from that file.

TODO: However if there is no context to load on the client side for the
new Nw Context Id, then the logic needs to handle this be setting up a
new empty client side context for this Nw Context Id. UPDATE: No cant
create a new empty client side context, because the issue of where to
write the recieved data can't be answered in a universal manner.