Client: Handle client side unknown nw context id
If in the middle of a nw test / transfer session, the server is
stopped and restarted with a new nw context id. And inturn if client
is not able to find a existing client side context for this new nw
context id, then the client will now go into a STATE_ERROR, in which
all recieved packets of this nw context id, will be skipped / dropped.

This is bcas, we can't safely identify a valid data file to write to,
if a client side context file is not found for a new nw context id,
seen suddenly in the middle of test / transfer involving some other
nw context id.

With the addition of this logic, now the CTXTAUTOLOAD option / define
is enabled by default on the client side.

END USER usage:

EIther start client freshly for each new nw context id, and force
quit it, so that it generates a corresponding client side context
file for that nw context id along with all relavent context like
nwContextId+ContextFileBase+DataFile+LostPktsInfo+...

OR

Manually create client side context files for all required nw context
ids, in turn containing appropriate context info for the client side.

Once either of the above two is done, then one can keep switching
between the different nw context ids, from the server side as required,
without worrying about restarting the client, each time the nw context
id and inturn data sent is changed as required on the server side.

TODO: However if a transfer is fully over for a given nw context id
from the client perspective and inturn in future if a totally new
unrelated transfer has to be done reusing the same nw context id,
then the client side context file requires to be cleared / updated
as required. Else client will ignore  (in future, once data from
already recieved packets are not processed the 2nd or subsequent
times) or intermix the data from the two unrelated test / transfer
which reuse the same nw context id.