Server: PI time4Clients
Renamed the PITotalTimeSecs to PITime4Clients, so that the name
represents the associated semantics.

This is now maintained as part of the network module.

One can specify the number of attempts and time to give to clients
as part of each attempt as part of presence_info call. If these
are not specified, then attempts default to 2 and time4Clients
default to 2 minutes.