Server:PI: Make it aggressive, if no known clients given
Previously only if a list of known clients is given, the pi logic
would keep attempting to get pi ack from them upto the specified
max attempts. Otherwise, i.e if no clients are specified, it will
only send 1 PI req and wait for the specified time for clients,
to see if any clients are there.

However now, if no clients are specified, then it will always try
to find if any clients are out there, for upto the maximum specified
attempts.