Server: NwContext Version handled --ncver
Now one can specify a version for the Nw Context (id) being transfered (inturn
tested), by using the --ncver on the server side.

This version is inturn sent as the 3rd entity in all the network packets sent
from server to the client. Previously this location was repeating the total
blocks involved as a dummy place holder.