Client: NwCtxtVer handled
If the recieved NwCtxtVer is different from current version for the
given NwCtxtId, then the logic will assume that the server / user
wants to send a new version of the NwCtxt test / transfer content.
So it will update the snm context such that the new content can be
recvd successfully subsequently.

Instead of the recieved NwCtxtVer matchs the current version for the
given NwCtxtId, then the logic continues the session associated with
that NwCtxtId.

As part of supporting the same, a additional debug assertion logic
added to the LinkedListRange wrt ll_free.