Server: Flags and PI send
The 3rd 4byte value in the packet has 8bit Flag at the MSB and 24bit Context
Version in the lower 3 bytes.

In turn the network module's presence info send helper func takes the current
value of the CurFlag variable and or's it with CtxtVer, after both have been
masked as required. So any code which wants to send a flag to the client
should first set network.CurFlag to the required value and inturn call
presence_info function. And once done (ie normally when PI function returns),
the CurFlag should be reset to the value it had before replacing it.

The nw-send-mcast logic inturn sets the CurFlag to the value which indicates
to the clients that they should save their context, just before sending the
in-between-the-transfer PIReq packets. And immidiately after that, it reverts
the CurFlag to a safe default value, which was previously saved.