Server:PI:Status logging update
As send_pireq is no longer called directly, but only thro the pi logic
which is also now moved into network module. So status logging wrt
sending of PIReq is now handled in the higher level pi logic rather
than the lower level send_pireq logic. This allows us to specify the
total number of times pi will be attempted and for how much time
we will wait for clients to respond, also in the status log