Server:send-mcast: Handle PI req timing
Now that I explicitly call presence info before the actual mcast data
packets are sent. THere is no need to sneak in, at the first attempt itself, a
presence info handshake in the logic which checks for periodic pi handshaking
with clients in the middle of the mcast data sends.

So I initialise the prevPITime to the time at the beginning of the mcast data
send logic rather than the previous 0.0.

NOTE: The initial PI phase before the actual data transfer begins, also helps
build the list of known clients. All subsequent PI handshakes work against this
known clients list. At the same time, new clients can join the transfer in the
middle and they will get added to the known clients list in the immidiate next
PI handshake (and inturn in the remaining subsequent PI handshakes even it will
be tracked as to whether it responds back or not).

Added a Makefile to server folder, so that a tag file can be generated for easy
navigation between the source codes. The resultant tag file is added to
.gitignore.

TODO: Need to add logic to save known clients list into the send-mcast context
file, as well as logic to load the same from the context file.