Server:_pi: if all known clients ackd then return
The pi code till now, was such that it will always wait for the specified
time4Clients, in a given attempt. After that it will check if it has recieved
ack from all known clients. In turn if it is in the KNOWN clients mode and all
known clients have acked, then it will return. Else it will try again, provides
there are more attempts remaining.

However now I have changed the logic such that, if one is in the KNOWN clients
mode, then as soon as it has got atleast one ack from all the known clients, it
will return immidiately (rather the check occurs once every 10 secs for now),
rather than waiting for the full time4Clients to expire (which is unnecessary
in KNOWN clients mode, if all clients have ackd).