Server:PI:Better logging wrt cur % loss
Now one can optionally pass the total number of blocks sent to clients till
now, to the PI logic. If not specified it defaults to the total number of
blocks involved in the current test/transfer.

In turn the PI logic will print the % of rcvd (indirectly percentage of losses)
wrt the current transfered amount.

Before the data transfer begins and at the end once all data is transfered,
there is not much use, wrt this new info which is printed/logged. However
during the middle of the transfer in mcast mode, this info helps get new
insight.