Server: Make it Cython compilable
Had to add a cython directive as a comment to status.py file so
that the print in it is treated as python3 related print with
named arguments, as I am using file= named argument to print.

Have added Makefile commands ccompile and cclean to generate
cython based version of the program.

Added clean which removes the __pycache__ folder