Server: select based sleep finetuning
As I am stupidly trying to read stdin (which can lead to a complex
interactions in normal case as well as in odd cases like putting
the process to background etc), I am now trapping the EOFError
exception which can be raised by input.

TODO: Need to avoid reading stdin, which is unneeded for select
based sleep to do its job.

NOTE: Maybe originally I wanted this to be a way to pause the
process in the middle and then do things based on input recieved.
However I didn't come around to doing it.