Server:nw-send-mcast: resumed test/transfer pi data
When I moved the pi logic into a func in network module and inturn
called it from mcast logic at the begin, in between and then at the
end. I had wrongly specified the pktid wrt start pi handshake as 0
instead of picking the proper value that will be there in pktid var
which gets suitably set if a test/transfer is being resumed.

THis oversight has been fixed now.