snm-status-analyse: devs2check list py file arg, testDevs2Check in release
Now the user can pass the python file which contains the devs2check
list with the names of all the devices which should be active in the
network, as a argument to the program.

Also the test python file with devs2check list in it, is now copied
as part of the server binary release. It is also commited to git now.

Scripts __pychache__ added to .gitignore.