run-nw-send-mcast helper does delayed start & dynamic ncver
To help the mcast based nw test logic to be run continuously in a
deployment, I have made the logic wait for 15 minutes, before
starting the actual logic. Inturn this is to be auto started during
boot time, by using the service start mechanisms like systemd unit
file etc.

Also as the program could be auto restarted by the system's init
daemon, if the system is on for a long time, I have modified the
logic to use the current hour as the ncver passed to the network
transfer. Thus as chances are the previous invocation and the
current invocation in the same power-on session on the server will
have two different hours, even thou the client wouldn't have been
restarted, it will still treat it as a new instance of the same
network transfer context.